# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Homelab Management API
  version: 1.0.0
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: string
paths:
  /auth/setup:
    post:
      summary: Create first admin account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, minLength: 3, maxLength: 64 }
                password: { type: string, minLength: 8, maxLength: 128 }
      responses:
        '201': { description: Setup complete }
        '422': { description: Invalid input }
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200': { description: Logged in }
        '401': { description: Invalid credentials }
        '422': { description: Invalid input }
  /auth/refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200': { description: Refreshed }
        '401': { description: Invalid refresh token }
  /auth/logout:
    post:
      security:
        - bearerAuth: []
      summary: Logout and revoke refresh token
      responses:
        '200': { description: Logged out }
  /services/status:
    get:
      security:
        - bearerAuth: []
      summary: Get service status map
      responses:
        '200': { description: Service status }
  /services/{name}/start:
    post:
      security:
        - bearerAuth: []
      summary: Start a service
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200': { description: Start requested }
        '400': { description: Invalid service name }
        '422': { description: Validation error }
  /services/{name}/stop:
    post:
      security:
        - bearerAuth: []
      summary: Stop a service
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200': { description: Stop requested }
        '400': { description: Invalid service name }
        '422': { description: Validation error }
  /settings/cloudflare-token:
    get:
      security:
        - bearerAuth: []
      summary: Get Cloudflare token status
      responses:
        '200': { description: Current config }
    put:
      security:
        - bearerAuth: []
      summary: Save Cloudflare token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, minLength: 20, maxLength: 4096 }
      responses:
        '200': { description: Token saved }
        '422': { description: Validation error }
  /settings/cloudflare-token/test:
    post:
      security:
        - bearerAuth: []
      summary: Test Cloudflare token
      responses:
        '200': { description: Test successful }
        '400': { description: Invalid token or test failed }
  /health:
    get:
      security:
        - bearerAuth: []
      summary: Get health status and thresholds
      responses:
        '200': { description: Health payload }
  /health/thresholds:
    put:
      security:
        - bearerAuth: []
      summary: Update health thresholds
      responses:
        '200': { description: Updated }
        '422': { description: Validation error }
  /audit-logs:
    get:
      security:
        - bearerAuth: []
      summary: List audit logs
      responses:
        '200': { description: Audit log page }
  /users:
    get:
      security:
        - bearerAuth: []
      summary: List admin accounts
      responses:
        '200': { description: List of users }
    post:
      security:
        - bearerAuth: []
      summary: Create a new admin account
      responses:
        '201': { description: User created }
        '409': { description: Username already exists }
        '422': { description: Validation error }
  /users/{id}/password:
    put:
      security:
        - bearerAuth: []
      summary: Reset an admin account's password
      responses:
        '200': { description: Password updated }
        '404': { description: User not found }
        '422': { description: Validation error }
  /users/{id}:
    delete:
      security:
        - bearerAuth: []
      summary: Delete an admin account
      responses:
        '200': { description: User deleted }
        '400': { description: Cannot delete own account or last remaining admin }
        '404': { description: User not found }
  /settings/exposure:
    get:
      security:
        - bearerAuth: []
      summary: Read first-start exposure provisioning settings (base domain, NPM, Cloudflare)
      responses:
        '200': { description: Exposure settings }
    put:
      security:
        - bearerAuth: []
      summary: Save first-start exposure provisioning settings
      responses:
        '200': { description: Saved }
        '422': { description: Validation error }
  /services/{name}/exposure:
    get:
      security:
        - bearerAuth: []
      summary: Read per-service exposure configuration and provisioning status
      responses:
        '200': { description: Exposure configuration }
        '400': { description: Invalid service name }
    put:
      security:
        - bearerAuth: []
      summary: Configure per-service exposure (enable/disable, upstream host/port)
      responses:
        '200': { description: Saved }
        '400': { description: Invalid service name }
        '422': { description: Validation error }
  /recovery/status:
    get:
      summary: Check recovery mode status
      responses:
        '200': { description: Recovery status }
  /recovery/enable:
    post:
      summary: Enable recovery mode (localhost only)
      responses:
        '200': { description: Enabled }
        '403': { description: Not localhost }
  /recovery/reset-admin-password:
    post:
      summary: Reset admin password (localhost + recovery mode)
      responses:
        '200': { description: Password reset }
        '422': { description: Validation error }